+++ includes/update.php
@@ -298,6 +298,9 @@
 
                 // From 4.14.1
                 'feed_cache_enabled' => 0,
+
+                // From [*next-version*]
+                'schedule_future_items' => 0,
 			)
 		);